Refactor and fix many warnings.

+++ core/src/org/rexellentgames/dungeon/Dungeon.java
@@ -103,7 +103,17 @@ public class Dungeon extends ApplicationAdapter {
 
 		File file = Gdx.files.external(SaveManager.getDir()).file();
 
-		for (File f : file.listFiles()) {
+		if (file == null) {
+			return;
+		}
+		
+		File[] files = file.listFiles();
+		
+		if (files == null) {
+			return;
+		}
+		
+		for (File f : files) {
 			f.delete();
 		}
 
@@ -255,7 +265,7 @@ public class Dungeon extends ApplicationAdapter {
 		
 		updateMouse(dt);
 
-		Gdx.gl.glClearColor(this.background.r, this.background.g, this.background.b, 1);
+		Gdx.gl.glClearColor(background.r, background.g, background.b, 1);
 		Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT | GL20.GL_DEPTH_BUFFER_BIT | (Gdx.graphics.getBufferFormat().coverageSampling ? GL20.GL_COVERAGE_BUFFER_BIT_NV : 0));
 
 		renderGame();
